Make types Send and Sync where applicable
All operations in OpenCL, except a few on kernels are thread-safe, see
https://www.khronos.org/registry/OpenCL/specs/3.0-unified/html/OpenCL_API.html#_multiple_host_threads

Hence it's safe to implement `Send` and `Sync` for all of them, except for
`Kernel`. There only `Send` is implemented as it is still safe to move it
between threads, only concurrent access is a problem.

Currently the shared virtual memory types don't implement `Sync`, although
they theoretically could. The reason is that they currently have in internal
reference to the `Context` type, which contains `Kernel`s, which are not
`Sync`. This could be changed by only storing the `cl_context` instead of
a full `Context` reference within the shared virtual memory types.